video importing ????

+++ asset/core.js
@@ -1,7 +1,7 @@
 const loadPost = require("../request/post_body");
 const header = process.env.XML_HEADER;
 const formidable = require("formidable");
-const fUtil = require("../fileUtil");
+const fUtil = require("../misc/file");
 const nodezip = require("node-zip");
 const base = Buffer.alloc(1, 0);
 const asset = require("./main");
@@ -10,6 +10,10 @@ const fs = require("fs");
 const starter = require("../starter/main");
 const util = require("../misc/util");
 const mp3Duration = require("mp3-duration");
+const tempfile = require("tempfile");
+const ffmpeg = require("fluent-ffmpeg");
+ffmpeg.setFfmpegPath(require("@ffmpeg-installer/ffmpeg").path);
+ffmpeg.setFfprobePath(require("@ffprobe-installer/ffprobe").path);
 function movieXml(v) {
 	const title = fs.readFileSync(process.env.META_FOLDER + `/${v.id}-title.txt`, 'utf8');
 	const tag = fs.readFileSync(process.env.META_FOLDER + `/${v.id}-tag.txt`, 'utf8');
@@ -26,6 +30,16 @@ function propXml(v) {
 	}" enable="Y" holdable="${meta.holdable}" headable="${meta.headable}" wearable="${meta.wearable}" placeable="${
 		meta.placeable
 	}" facing="left" width="0" height="0" asset_url="/assets/${v.type}/${v.id}.${ext}"/>`;
+}
+function videoXml(v) {
+	const title = fs.readFileSync(process.env.META_FOLDER + `/${v.id}-title.txt`, 'utf8');
+	const ext = fs.existsSync(process.env.META_FOLDER + `/${v.id}-ext.txt`, 'utf8') ? fs.readFileSync(process.env.META_FOLDER + `/${
+		v.id
+	}-ext.txt`, 'utf8') : getExt(title);
+	const meta = require('.' + process.env.META_FOLDER + `/${v.id}-meta.json`);
+	return `<prop subtype="video" id="${v.id}.${ext}" enc_asset_id="${v.id}.${ext}" name="${
+		title
+	}" enable="Y" placeable="1" facing="left" width="${meta.width}" height="${meta.height}" asset_url="/assets/${v.type}/${v.id}.${ext}" thumbnail_url="/assets/${v.type}/${v.id}.png"/>`;
 }
 function getExt(buffer) {
 	const dot = buffer.lastIndexOf(".");
@@ -71,6 +85,29 @@ async function listAssets(data) {
 		}	
 	}
 	return Buffer.from(xmlString);	
+}
+function convertVideoToFlv(ut, type, _ext, _buffer, subtype, filepath) {
+	const files = {};
+	const temppath = tempfile(".flv");
+	ffmpeg(filepath).ffprobe((e, data) => {
+		if (e) console.log(e);
+		const meta = {
+			width: data.streams[0].width,
+			height: data.streams[0].height
+		};
+
+		// convert the video to an flv
+		ffmpeg(filepath).output(temppath).on("end", async () => {
+			const buff = fs.readFileSync(temppath);
+			meta.file = `vi-${asset.saveStream(ut, type, "flv", buff, subtype)}`;
+			const vId = meta.file.slice(0, -4);
+			const id = Number.parseInt(vId.substr(2));
+			// save the first frame
+			ffmpeg(filepath).seek("0:00").output(fUtil.getFileIndexForAssets("video-", ".png", id)).outputOptions("-frames", "1").on("end", () => files[meta.file]).run();
+			fs.writeFileSync(process.env.META_FOLDER + `/${vId}-meta.json`, JSON.stringify(meta));
+			fs.writeFileSync(process.env.META_FOLDER + `/${vId}-title.txt`, meta.file);
+		}).on("error", (e) => console.log("Error converting video:", e)).run();
+	});
 }
 
 /**
@@ -161,6 +198,26 @@ module.exports = function (req, res, url) {
 									return;
 								});
 								break;
+							} case "prop": {
+								function listProps() {
+									return new Promise(async res => {
+										files = asset.list("video", "flv");
+										const xml = `${header}<ugc more="0">${files.map((v) => videoXml(v)).join("")}</ugc>`;
+										const zip = nodezip.create();
+										fUtil.addToZip(zip, "desc.xml", Buffer.from(xml));
+										res(await zip.zip());
+									});
+								}
+								listProps().then((buff) => {
+									res.setHeader("Content-Type", "application/zip");
+									res.write(base);
+									res.end(buff);
+								}).catch(e => {
+									res.end(1 + util.assetFail(e));
+									console.log(e);
+									return;
+								});
+								break;
 							}
 						}
 					}).catch(e => console.log(e));
@@ -182,7 +239,7 @@ module.exports = function (req, res, url) {
 						prefix = "";
 						const path = files.import.path;
 						const name = files.import.name;
-						const buffer = fs.readFileSync(path);
+						var buffer = fs.readFileSync(path);
 						switch (type) {
 							case "vo":
 								type = "sound";
@@ -199,32 +256,39 @@ module.exports = function (req, res, url) {
 								subtype = "bgmusic";
 								prefix = "bg";
 								break;
-						}
-						asset.save(ut, type, ext, buffer, subtype).then(id => {
-							asset.createMeta(id, name, type, subtype);
-							switch (type) {
-								case "prop": {
-									const meta = {
-										holdable: "0",
-										wearable: "0",
-										headable: "0",
-										placeable: "1"
-									};
-									fs.writeFileSync(process.env.META_FOLDER + `/p-${id}-meta.json`, JSON.stringify(meta));
-									break;
-								} case "sound": {
-									mp3Duration(buffer, (e, d) => {
-										if (e) {
-											console.log(e);
-											return;
-										}
-										fs.writeFileSync(process.env.META_FOLDER + `/${prefix}-${id}-dur.txt`, `dur.${d * 1e3}`);
-									});
-									break;
-								}
+							case "vi": {
+								type = "prop";
+								subtype = "video";
+								break;
 							}
-						}).catch(e => console.log(e));
-						fs.unlinkSync(path);
+						}
+						if (subtype != "video") {
+							asset.save(ut, type, ext, buffer, subtype).then(id => {
+								asset.createMeta(id, name, type, subtype);
+								switch (type) {
+									case "prop": {
+										const meta = {
+											holdable: "0",
+											wearable: "0",
+											headable: "0",
+											placeable: "1"
+										};
+										fs.writeFileSync(process.env.META_FOLDER + `/p-${id}-meta.json`, JSON.stringify(meta));
+										break;
+									} case "sound": {
+										mp3Duration(buffer, (e, d) => {
+											if (e) {
+												console.log(e);
+												return;
+											}
+											fs.writeFileSync(process.env.META_FOLDER + `/${prefix}-${id}-dur.txt`, `dur.${d * 1e3}`);
+										});
+										break;
+									}
+								}
+							}).catch(e => console.log(e));
+							fs.unlinkSync(path);
+						} else convertVideoToFlv(ut, type, ext, buffer, subtype, path);
 					});
 					return true;
 				} case "/goapi/deleteAsset/": {
